+++ .gitignore
@@ -1,2 +1,2 @@
-lib/
-dist/
+/lib/
+/dist/
